move the xored static methods to app.utility.base_world.BaseWorld
moved the tests back again to test_base_world.py
