+++ tests/utility/test_base_world.py
@@ -91,3 +91,51 @@ class TestBaseWorld:
     def test_is_base64(self):
         b64str = 'aGVsbG8gd29ybGQgZnJvbSB1bml0IHRlc3QgbGFuZAo='
         assert BaseWorld.is_base64(b64str)
+
+    def test_walk_file_path_exists_nonxor(self, loop, text_file):
+        ret = loop.run_until_complete(BaseWorld.walk_file_path(text_file.dirname, text_file.basename))
+        assert ret == text_file
+
+    def test_walk_file_path_notexists(self, loop, text_file):
+        ret = loop.run_until_complete(BaseWorld.walk_file_path(text_file.dirname, 'not-a-real.file'))
+        assert ret is None
+
+    def test_walk_file_path_xor_fn(self, loop, tmpdir):
+        f = tmpdir.mkdir('txt').join('xorfile.txt.xored')
+        f.write("test")
+        ret = loop.run_until_complete(BaseWorld.walk_file_path(f.dirname, 'xorfile.txt'))
+        assert ret == f
+
+    def test_remove_xored_extension(self):
+        test_value = 'example_file.exe.xored'
+        expected_value = 'example_file.exe'
+        ret = BaseWorld.remove_xored_extension(test_value)
+        assert ret == expected_value
+
+    def test_remove_xored_extension_to_non_xored_file(self):
+        test_value = 'example_file.exe'
+        expected_value = 'example_file.exe'
+        ret = BaseWorld.remove_xored_extension(test_value)
+        assert ret == expected_value
+
+    def test_add_xored_extension(self):
+        test_value = 'example_file.exe'
+        expected_value = 'example_file.exe.xored'
+        ret = BaseWorld.add_xored_extension(test_value)
+        assert ret == expected_value
+
+    def test_add_xored_extension_to_xored_file(self):
+        test_value = 'example_file.exe.xored'
+        expected_value = 'example_file.exe.xored'
+        ret = BaseWorld.add_xored_extension(test_value)
+        assert ret == expected_value
+
+    def test_is_extension_xored_true(self):
+        test_value = 'example_file.exe.xored'
+        ret = BaseWorld.is_extension_xored(test_value)
+        assert ret is True
+
+    def test_is_extension_xored_false(self):
+        test_value = 'example_file.exe'
+        ret = BaseWorld.is_extension_xored(test_value)
+        assert ret is False
